Generalize use of KEYMAP in evil-define-key
In evil-define-key, allow keymap to be 'global, 'local or some other quoted
symbol, which is assumed to be a minor mode.

Using 'global, corresponds to making a global binding in the corresponding
global evil keymap (similar to evil-global-set-key). Using 'local, corresponds
to making a local binding in the corresponding local evil keymap (similar to
evil-local-set-key).

Any other quoted symbol is assumed to be the name of a minor mode, and the
binding is dispatched to evil-define-minor-mode-key.

Make evil-define-key* support global and local symbols.

Allow a list of states in evil-define-minor-mode-key.

This makes evil-define-key serve as a general keybinding function.